Fixed various errors in deprecated concepts
Note that there are still a) deprecated concepts without a 'consider' or 'replaced_by', and b) valid concepts with 'consider'. Both may make sense; in case of a) if there is nothing relevantly related in scope of EDAM, and in case of b) if there is concept that is often a better choice. Note, however, that b) cases should thus be perhaps considered for deprecation.