Fixed Data management vs. Biological databases bug
This was a major "bug" introduced in 1.24 and pertaining in 1.25, which renamed topic_3071 from Data management to Biological databases, without keeping the original meaning.

(Note: Exploration of Bio.tools data at the moment indicates that probably >500 records were annotated with topic_3071 as Data management, and later <100 records with topic_3071 as Biological databases. Especially the latter have to be manually curated! See lists in https://github.com/bio-tools/content/pull/521)

- The current conclusion is that 'Biological databases' is not needed as a topic (just like 'Bioinformatics tools', as it is the topic 'Biology' plus type of resource being a 'Data resource' or related).
- **Data management** is now topic_3071, as it was from version _beta13_ until _1.23_ inclusively.
- **Data rescue** has been added as topic_4011. (Note that topic_4011 was shortly 'Data management' in the unstable _1.26_dev_ development version. In the very unlikely case that this was used anywhere, some manual curation would be needed.)