Recommended a shallow clone in CONTRIBUTING.md
This is not super relevant yet, when it comes to the size and download time. But it doesn't confuse the user with many branches and complex history.

I assume it will be useful when it comes to size and time in the future, when we manage to remove all the released "binaries".